+++ superset/assets/javascripts/explorev2/components/Filters.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 // import { Tab, Row, Col, Nav, NavItem } from 'react-bootstrap';
-import Select from 'react-select';
+import Filter from './Filter';
 import { Button } from 'react-bootstrap';
 import { connect } from 'react-redux';
 import { bindActionCreators } from 'redux';
@@ -21,28 +21,6 @@ const defaultProps = {
 };
 
 class Filters extends React.Component {
-  constructor(props) {
-    super(props);
-    const opChoices = this.props.prefix === 'flt' ?
-      ['in', 'not in'] : ['==', '!=', '>', '<', '>=', '<='];
-    this.state = {
-      opChoices,
-    };
-  }
-  changeCol(filter, colOpt) {
-    const val = (colOpt) ? colOpt.value : null;
-    this.props.actions.changeFilter(filter, 'col', val);
-  }
-  changeOp(filter, opOpt) {
-    const val = (opOpt) ? opOpt.value : null;
-    this.props.actions.changeFilter(filter, 'op', val);
-  }
-  changeValue(filter, event) {
-    this.props.actions.changeFilter(filter, 'value', event.target.value);
-  }
-  removeFilter(filter) {
-    this.props.actions.removeFilter(filter);
-  }
   addFilter() {
     this.props.actions.addFilter({
       id: shortid.generate(),
@@ -58,49 +36,12 @@ class Filters extends React.Component {
       // only display filters with current prefix
       if (filter.prefix === this.props.prefix) {
         filters.push(
-          <div>
-            <div className="row space-1">
-              <Select
-                className="col-lg-12"
-                multi={false}
-                name="select-column"
-                placeholder="Select column"
-                options={this.props.filterColumnOpts.map((o) => ({ value: o, label: o }))}
-                value={filter.col}
-                autosize={false}
-                onChange={this.changeCol.bind(this, filter)}
-              />
-            </div>
-            <div className="row space-1">
-              <Select
-                className="col-lg-4"
-                multi={false}
-                name="select-op"
-                placeholder="Select operator"
-                options={this.state.opChoices.map((o) => ({ value: o, label: o }))}
-                value={filter.op}
-                autosize={false}
-                onChange={this.changeOp.bind(this, filter)}
-              />
-              <div className="col-lg-6">
-                <input
-                  type="text"
-                  onChange={this.changeValue.bind(this, filter)}
-                  value={filter.value}
-                  className="form-control input-sm"
-                  placeholder="Filter value"
-                />
-              </div>
-              <div className="col-lg-2">
-                <Button
-                  bsSize="small"
-                  onClick={this.removeFilter.bind(this, filter)}
-                >
-                  <i className="fa fa-minus" />
-                </Button>
-              </div>
-            </div>
-          </div>
+          <Filter
+            filterColumnOpts={this.props.filterColumnOpts}
+            actions={this.props.actions}
+            prefix={this.props.prefix}
+            filter={filter}
+          />
         );
       }
     });
